refactor: Check if ghq is already installed

+++ ghq/install.sh
@@ -10,18 +10,21 @@ source "$DOTFILES_ROOT/lib/logger.sh"
 
 info 'Installing ghq'
 
-if ! command -v brew > /dev/null; then
-  user 'Homebrew not found. Would you like to install Homebrew first? (y/n)'
-  read -r -p '> ' install_brew
+if ! command -v ghq > /dev/null; then
+  if ! command -v brew > /dev/null; then
+    user 'Homebrew not found. Would you like to install Homebrew first? (y/n)'
+    read -r -p '> ' install_brew
 
-  if [[ "$install_brew" =~ ^[Yy]$ ]]; then
-    source "$DOTFILES_ROOT/homebrew/install.sh"
-    source "$DOTFILES_ROOT/homebrew/path.zsh"
-  else
-    fail 'Homebrew is required to install ghq.'
+    if [[ "$install_brew" =~ ^[Yy]$ ]]; then
+      source "$DOTFILES_ROOT/homebrew/install.sh"
+      source "$DOTFILES_ROOT/homebrew/path.zsh"
+    else
+      fail 'Homebrew is required to install ghq.'
+    fi
   fi
+  brew install ghq
+
+  success 'ghq installed'
+else
+  success 'ghq is already installed'
 fi
-
-brew install ghq
-
-success 'ghq installed'
